refactoring some code and cleared unused function
