+++ client/actions/search.js
@@ -1,18 +1,10 @@
-import { searchAPI, searchMaoriRegionAPI } from '../apis/search'
+import { searchAPI } from '../apis/search'
 
 export const SEARCH_REGION = 'SEARCH_REGION'
-export const MAORI_SEARCH = 'MAORI_SEARCH'
 
 export function searchRegion (region) {
   return {
     type: SEARCH_REGION,
-    payload: region
-  }
-}
-
-export function searchMaoriRegion (region) {
-  return {
-    type: MAORI_SEARCH,
     payload: region
   }
 }
@@ -28,14 +20,3 @@ export function fetchSearchedRegions (data) {
   }
 }
 
-export function fetchSearchedMaoriRegions (data) {
-  return async (dispatch) => {
-    try {
-    const result = await searchMaoriRegionAPI(data)
-    dispatch(searchMaoriRegion(result))
-  } catch (err) {
-    console.error(err.message)
-  }
-
-}
-}
