--- client/components/Image.tsx
@@ -1,47 +1,0 @@
-import React, { useState } from 'react';
-import ImageUploading from 'react-images-uploading';
-
-export function App() {
-  const [images, setImages] = useState([]);
-
-  const onChange = (imageList) => {
-    console.log(imageList);
-    setImages(imageList);
-  };
-
-  return (
-    <div className="App">
-      <ImageUploading
-        value={images}
-        onChange={onChange}
-        dataURLKey="data_url"
-      >
-        {({
-          imageList,
-          onImageUpload,
-          onImageRemoveAll,
-          isDragging,
-          dragProps,
-        }) => (
-          <div className="upload__image-wrapper">
-            <button
-              style={isDragging ? { color: 'red' } : undefined}
-              onClick={onImageUpload}
-              {...dragProps}
-            >
-              Click or Drop here
-            </button>
-            &nbsp;
-            <button onClick={onImageRemoveAll}>Remove all images</button>
-            {imageList.map((image, index) => (
-              <div key={index} className="image-item">
-                <img src={image['data_url']} alt="" width="100" />
-           
-              </div>
-            ))}
-          </div>
-        )}
-      </ImageUploading>
-    </div>
-  );
-}
