+++ client/components/StoryCopy.tsx
@@ -5,9 +5,8 @@ import { useSelector, useDispatch } from 'react-redux'
 import Map from './Map'
 
 import { fetchOneStory } from '../actions/stories'
-import { deleteStoryAPI } from '../apis/story'
+import { deleteStoryAPI } from '../apis/stories'
 import { fetchAddSavedStory } from '../actions/savedStories'
-import { addSavedStoryAPI } from '../apis/savedStories'
 
 function StoryCopy() {
   const { id } = useParams()
