+++ client/reducers/index.js
@@ -1,21 +1,17 @@
 import { combineReducers } from 'redux'
 
+import map from './map'
+import user from './loggedInUser'
 import search from './search'
 import stories from './stories'
-import user from './loggedInUser'
 import regions from './regions'
-import activepage from './activepage'
-import maoriSearch from './maoriSearch'
-import map from './map'
 import savedStories from './savedStories'
 
 export default combineReducers({
-  user,
-  stories,
-  search,
-  activepage,
-  regions,
-  maoriSearch,
   map,
+  user,
+  search,
+  stories,
+  regions,
   savedStories
 })
