+++ server/db/regions.js
@@ -1,11 +1,8 @@
 const conn = require('./connection')
 
 function getRegions(db = conn) {
-  return db('regions')  
-  .select()  
+  return db('regions')  .select()  
 
 }
 
-module.exports = {
-  getRegions
-}
+module.exports = getRegions
